requirements.txt
# Core Dependencies
langchain==0.1.20
langgraph==0.0.55
langchain-groq==0.1.3
langchain-community==0.0.38

# ML/Embeddings (CPU-only)
sentence-transformers==2.5.1
torch==2.2.1
faiss-cpu==1.8.0

# Graph/Ontology
networkx==3.2.1

# API/Web
fastapi==0.110.0
uvicorn==0.27.1
streamlit==1.32.0
pydantic==2.6.3
python-dotenv==1.0.1

# Data Generation/Handling
faker==24.0.0

# LLMOps/Monitoring
langsmith==0.1.17

# Testing/Quality
pytest==8.1.1
flake8==7.0.0

# Utilities
tenacity==8.2.3
numpy==1.26.4
pandas==2.2.1